解决中文决策树测试时的编码问题，返回结果也是中文
